Move shared capybara config into shared helper
